Restructure project
Move tests to a separate folder
