+++ seminar7/service/singer.py
@@ -8,6 +8,13 @@ class SingerService:
         self.__validator = SingerValidator()
 
     def add(self, singer_id, name, genre):
+        """
+        Add a new singer
+        :param singer_id: the id of the new singer
+        :param name: the name of the new singer
+        :param genre: the genre of the new singer
+        :return:
+        """
         singer = Singer(singer_id, name, genre)
         self.__validator.validate(singer)
         self.__repo.add(singer)
@@ -16,7 +23,10 @@ class SingerService:
         return self.__repo.get_all()
 
     def get_nr_singers_by_all_genres(self):
-        """"""
+        """
+        Returns a dictionary with the number of singers for each genre
+        :return:
+        """
         result = dict()
         for singer in self.get_all():
             if singer.get_genre() in result:
