+++ seminar7/tests/service.py
@@ -1,27 +1,8 @@
-from seminar7.domain import Singer, ValidatorException
-from seminar7.repository import SingerRepository
+from seminar7.domain import ValidatorException
 from seminar7.service import SingerService
 
 
-def test_singer():
-    singer = Singer(1, "Eminem", "rap")
-    assert singer.get_name() == "Eminem"
-    assert singer.get_id() == 1
-    assert singer.get_genre() == "rap"
-    assert str(singer) == "Singer(id=1,name=Eminem,genre=rap)"
-
-    assert singer != Singer(2, "Eminem", "rap")
-    assert singer == Singer(1, "Eminem", "rap")
-
-
-def test_repository():
-    repo = SingerRepository()
-    singer = Singer(1, "Eminem", "rap")
-    repo.add(singer)
-    assert len(repo) == 1
-
-
-def test_service():
+def test_singer_service():
     service = SingerService()
 
     try:
@@ -38,7 +19,7 @@ def test_service():
     assert len(service.get_all()) == 1
 
 
-def test_servive_genre_report():
+def test_performance_service_genre_report():
     service = SingerService()
     assert service.get_nr_singers_by_all_genres() == {}
     service.add(1, "Eminem", "rap")
@@ -54,7 +35,5 @@ def test_servive_genre_report():
 
 
 if __name__ == "__main__":
-    test_singer()
-    test_repository()
-    test_service()
-    test_servive_genre_report()
+    test_singer_service()
+    test_performance_service_genre_report()
